added some bugfixes

+++ src/utils/api/db/calls/sequences.ts
@@ -34,6 +34,8 @@ export const createSequenceCall = (supabaseClient: RelayDatabase) => async (inse
 };
 
 export const deleteSequenceCall = (supabaseClient: RelayDatabase) => async (id: string) => {
+    const { error: error_steps } = await supabaseClient.from('sequence_steps').delete().eq('sequence_id', id);
+    if (error_steps) throw error_steps;
     const { error } = await supabaseClient.from('sequences').delete().eq('id', id);
     if (error) throw error;
 };
